+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Made --no-zip enable --no-cleanup by default
 - NIF processing logs will not show the block id of shapes
 - getFile now logs which BSA it's pulling a file from
+- Fixed crash that would occur if a shape doesn't have a shader
 
 ## [0.2.0] - 2024-06-04